Mv TC utils module into src
